Use two variables instead of one for separate FF instances

+++ functions/helloworld/test/sample.integration.pubsub.test.js
@@ -22,12 +22,6 @@ const waitPort = require('wait-port');
 // [END functions_pubsub_integration_test]
 
 describe('functions_helloworld_pubsub integration test', () => {
-  let ffProc;
-  afterEach(() => {
-    if (ffProc) {
-      ffProc.kill();
-    }
-  });
   // [START functions_pubsub_integration_test]
   it('helloPubSub: should print a name', async () => {
     const name = uuid.v4();
@@ -35,7 +29,7 @@ describe('functions_helloworld_pubsub integration test', () => {
 
     const encodedName = Buffer.from(name).toString('base64');
     const pubsubMessage = {data: {data: encodedName}};
-    ffProc = spawn('npx', [
+    const ffProc = spawn('npx', [
       'functions-framework',
       '--target',
       'helloPubSub',
@@ -44,43 +38,51 @@ describe('functions_helloworld_pubsub integration test', () => {
       '--port',
       PORT,
     ]);
-    const ffProcHandler = new Promise((resolve, reject) => {
-      let stdout = '';
-      let stderr = '';
-      ffProc.stdout.on('data', data => (stdout += data));
-      ffProc.stderr.on('data', data => (stderr += data));
-      ffProc.on('exit', code => {
-        if (code === 0) {
-          resolve(stdout);
-        } else {
-          stderr = `Error code: ${code}\n${stderr}`;
-          reject(stderr);
-        }
+
+    try {
+      const ffProcHandler = new Promise((resolve, reject) => {
+        let stdout = '';
+        let stderr = '';
+        ffProc.stdout.on('data', data => (stdout += data));
+        ffProc.stderr.on('data', data => (stderr += data));
+        ffProc.on('exit', code => {
+          if (code === 0) {
+            resolve(stdout);
+          } else {
+            stderr = `Error code: ${code}\n${stderr}`;
+            reject(stderr);
+          }
+        });
       });
-    });
-    await waitPort({host: 'localhost', port: PORT});
+      await waitPort({host: 'localhost', port: PORT});
 
-    // Send HTTP request simulating Pub/Sub message
-    // (GCF translates Pub/Sub messages to HTTP requests internally)
-    const response = await request({
-      url: `http://localhost:${PORT}/`,
-      method: 'POST',
-      data: pubsubMessage,
-    });
-    ffProc.kill();
+      // Send HTTP request simulating Pub/Sub message
+      // (GCF translates Pub/Sub messages to HTTP requests internally)
+      const response = await request({
+        url: `http://localhost:${PORT}/`,
+        method: 'POST',
+        data: pubsubMessage,
+      });
+      ffProc.kill();
 
-    assert.strictEqual(response.status, 204);
+      assert.strictEqual(response.status, 204);
 
-    // Wait for the functions framework to stop
-    const stdout = await ffProcHandler;
-    assert.match(stdout, new RegExp(`Hello, ${name}!`));
+      // Wait for the functions framework to stop
+      const stdout = await ffProcHandler;
+      assert.match(stdout, new RegExp(`Hello, ${name}!`));
+    } catch {
+      if (ffProc) {
+        // Make sure the functions framework is stopped
+        ffProc.kill();
+      }
+    }
   });
   // [END functions_pubsub_integration_test]
 
   it('helloPubSub: should print hello world', async () => {
     const pubsubMessage = {data: {}};
     const PORT = 8089; // Each running framework instance needs a unique port
-    ffProc = spawn('npx', [
+    const ffProc = spawn('npx', [
       'functions-framework',
       '--target',
       'helloPubSub',
@@ -89,30 +91,37 @@ describe('functions_helloworld_pubsub integration test', () => {
       '--port',
       PORT,
     ]);
-    const ffProcHandler = new Promise((resolve, reject) => {
-      let stdout = '';
-      let stderr = '';
-      ffProc.stdout.on('data', data => (stdout += data));
-      ffProc.stderr.on('data', data => (stderr += data));
-      ffProc.on('error', reject);
-      ffProc.on('exit', c => (c === 0 ? resolve(stdout) : reject(stderr)));
-    });
-    await waitPort({host: 'localhost', port: PORT});
+    try {
+      const ffProcHandler = new Promise((resolve, reject) => {
+        let stdout = '';
+        let stderr = '';
+        ffProc.stdout.on('data', data => (stdout += data));
+        ffProc.stderr.on('data', data => (stderr += data));
+        ffProc.on('error', reject);
+        ffProc.on('exit', c => (c === 0 ? resolve(stdout) : reject(stderr)));
+      });
+      await waitPort({host: 'localhost', port: PORT});
 
-    // Send HTTP request simulating Pub/Sub message
-    // (GCF translates Pub/Sub messages to HTTP requests internally)
-    const response = await request({
-      url: `http://localhost:${PORT}/`,
-      method: 'POST',
-      data: pubsubMessage,
-    });
-    ffProc.kill();
+      // Send HTTP request simulating Pub/Sub message
+      // (GCF translates Pub/Sub messages to HTTP requests internally)
+      const response = await request({
+        url: `http://localhost:${PORT}/`,
+        method: 'POST',
+        data: pubsubMessage,
+      });
+      ffProc.kill();
 
-    assert.strictEqual(response.status, 204);
+      assert.strictEqual(response.status, 204);
 
-    // Wait for functions-framework process to exit
-    const stdout = await ffProcHandler;
-    assert.match(stdout, /Hello, World!/);
+      // Wait for functions-framework process to exit
+      const stdout = await ffProcHandler;
+      assert.match(stdout, /Hello, World!/);
+    } catch {
+      if (ffProc) {
+        // Make sure the functions framework is stopped
+        ffProc.kill();
+      }
+    }
   });
   // [START functions_pubsub_integration_test]
 });
